Python interactive game
This is Python program is an interactive game with the user. The user is asked for their name, then what game they would like to play: math or binary.
Next, they are asked how many questions they would like in the game which is where the setter property is implemented for outside the range of 1-10.
At the end of each game, the user's score is printed and their score updated in the score file.